convert env file into package.env dir
required for crossdev to run (so far, it says)

also package.accept_keywords file to dir!

and some files were dups in the sense that the .conf file specified in 'env'
existed + also the dir too with its file, not sure whether only one of
them was used or both were merged.